monodepth/requirements.txt
# Requires Python >= 3.8
torch>=2.0.0
torchvision>=0.15.0
torchinfo>=1.8.0
numpy>=1.19.0
opencv-python>=4.10.0
matplotlib>=3.3.0
pillow>=8.0.0
scipy>=1.5.0
tqdm>=4.50.0